Restructure cpp files
